refactor: highlightText 컴포넌트로 변경
