+++ src/features/post/ui/PostDetail.tsx
@@ -1,8 +1,8 @@
 import { DialogContent, DialogHeader, DialogTitle } from "@shared/ui"
-import { highlightText } from "@pages/PostsManagerPage.tsx"
 import { Comments } from "@features/post/ui/Comments.tsx"
 import { Post } from "@entities/post/types"
 import { useSearchParams } from "react-router-dom"
+import { HighlightText } from "@shared/ui"
 
 interface PostDetailProps {
   post: Post
@@ -16,10 +16,14 @@ export function PostDetail(props: PostDetailProps) {
   return (
     <DialogContent className="max-w-3xl">
       <DialogHeader>
-        <DialogTitle>{highlightText(post?.title, searchQuery)}</DialogTitle>
+        <DialogTitle>
+          <HighlightText text={post?.title} highlight={searchQuery} />
+        </DialogTitle>
       </DialogHeader>
       <div className="space-y-4">
-        <p>{highlightText(post?.body, searchQuery)}</p>
+        <p>
+          <HighlightText text={post?.body} highlight={searchQuery} />
+        </p>
         <Comments postId={post.id} />
       </div>
     </DialogContent>
