+++ src/shared/ui/index.ts
@@ -3,3 +3,4 @@ export * from "./card"
 export * from "./dialog"
 export * from "./inputs"
 export * from "./table"
+export * from "./thypograph"
