+++ src/shared/ui/thypograph/index.ts
@@ -1,0 +1,1 @@
+export * from "./HighlightText.tsx"
